# doc/JJMasterData.Documentation/restapi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JJMasterData.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/accounts/login": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Realizes the login.",
        "requestBody": {
          "description": "user={login}&password={plain text}",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/AccountLogin"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountLogin"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountLogin"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AccountLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccessInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccessInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccessInfo"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/accounts/changepassword": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Change the user password.",
        "requestBody": {
          "description": "User data",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/AccountChange"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountChange"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountChange"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AccountChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccessInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccessInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccessInfo"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          },
          "406": {
            "description": "Not Acceptable"
          }
        }
      }
    },
    "/api/accounts/recoverpassword": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Recover the user password.",
        "requestBody": {
          "description": "user={Login}&appId={App Id}",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/AccountRecover"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountRecover"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountRecover"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AccountRecover"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccessInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccessInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccessInfo"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/masterApi/{elementName}/{id}": {
      "get": {
        "tags": [
          "MasterApi"
        ],
        "parameters": [
          {
            "name": "elementName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { }
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { }
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "MasterApi"
        ],
        "parameters": [
          {
            "name": "elementName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLetter"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLetter"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLetter"
                }
              }
            }
          }
        }
      }
    },
    "/masterApi/{elementName}": {
      "post": {
        "tags": [
          "MasterApi"
        ],
        "parameters": [
          {
            "name": "elementName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "replace",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": { }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": { }
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": { }
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": { }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLetter"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLetter"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLetter"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "MasterApi"
        ],
        "parameters": [
          {
            "name": "elementName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": { }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": { }
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": { }
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": { }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLetter"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLetter"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLetter"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "MasterApi"
        ],
        "parameters": [
          {
            "name": "elementName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": { }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": { }
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": { }
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": { }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLetter"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLetter"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLetter"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountChange": {
        "required": [
          "pwdConfirm",
          "pwdCurrent",
          "pwdNew",
          "user"
        ],
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "description": "User login"
          },
          "pwdCurrent": {
            "type": "string",
            "description": "Current Password"
          },
          "pwdNew": {
            "type": "string",
            "description": "New Password"
          },
          "pwdConfirm": {
            "type": "string",
            "description": "Confirm Password"
          }
        },
        "additionalProperties": false
      },
      "AccountLogin": {
        "required": [
          "password",
          "user"
        ],
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "description": "User login"
          },
          "password": {
            "type": "string",
            "description": "Current password without cryptography"
          },
          "appId": {
            "type": "string",
            "description": "AppId number from smartphone,  \r\nit's used on recover password",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AccountRecover": {
        "required": [
          "user"
        ],
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "description": "User login"
          },
          "appId": {
            "type": "string",
            "description": "AppId number from smartphone",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ResponseLetter": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "validationList": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          },
          "data": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UserAccessInfo": {
        "type": "object",
        "properties": {
          "errorId": {
            "type": "integer",
            "description": "Código do erro:\r\n<br />0   = OK\r\n<br />100 = Unexpected error\r\n<br />101 = Password sent by email if existing\r\n<br />110 = Invalid Access\r\n<br />112 = Sua conta excedeu o limite de tentitivas\r\n<br />113 = Blocked user\r\n<br />130 = Password change required\r\n<br />131 = Your password will expire soon, change it as soon as possible\r\n<br />301 = Password cannot be blank!\r\n<br />302 = Password cannot contain user!\r\n<br />303 = Password must contain at least one number!\r\n<br />304 = The password must contain at least one uppercase character!\r\n<br />305 = The password contains at least one of the following characters: !@#$%^*()_+&<br />306 = Password cannot contain a colon(:)\r\n<br />307 = Password and confirmation password are not the same!\r\n<br />308 = Your current password is incorrect! It will not be possible to change to the new password.",
            "format": "int32"
          },
          "isValid": {
            "type": "boolean",
            "description": "Retorna verdadeiro se o login foi realizado com sucesso"
          },
          "message": {
            "type": "string",
            "description": "Descrição do erro",
            "nullable": true
          },
          "token": {
            "type": "string",
            "description": "Token a ser utilizado nas requisições",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "description": "Id do Usuário",
            "nullable": true
          },
          "version": {
            "type": "string",
            "description": "Versão do sistema",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Informações do login do usuário"
      },
...
    },
    "securitySchemes": {
      "token": {
        "type": "apiKey",
        "description": "API Key Authentication",
        "name": "token",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "token": [ ]
    }
  ]
}